Address Noah's comments
